// src/views/Orders/Chemicals/interfaces.ts
export interface ModalTypes {
  LABRECETEID?: number | string | any;
  LABRECETEKODU?: string
  }

  export interface IModalForm {
    LABRECETEADI: string
        
    LABRECETEKODU: string
    LABRECETEID: number | string

    FIRMAID: string
    PANTONEKODU: string
    DOVIZID: string
    KULLANICIADI: string
    ACIKLAMA: string
    ESKILABRECETEKODU: string
    RENKDERINLIGIADI: string
    RECETETURAADI: string
    TALEPTARIHI: string
  } 